+++ ra-optimisation/rao-commons/src/main/java/com/farao_community/farao/rao_commons/linear_optimisation/fillers/MaxMinRelativeMarginFiller.java
@@ -19,7 +19,6 @@ import com.google.ortools.linearsolver.MPObjective;
 import com.google.ortools.linearsolver.MPVariable;
 
 import java.util.Optional;
-import java.util.Set;
 
 import static com.farao_community.farao.commons.Unit.MEGAWATT;
 
@@ -31,8 +30,8 @@ public class MaxMinRelativeMarginFiller extends MaxMinMarginFiller {
     private double negativeMarginObjectiveCoefficient;
     private double ptdfSumLowerBound;
 
-    public MaxMinRelativeMarginFiller(Unit unit, double pstPenaltyCost, Set<String> operatorsNotToOptimize, double negativeMarginObjectiveCoefficient, double ptdfSumLowerBound) {
-        super(unit, pstPenaltyCost, operatorsNotToOptimize);
+    public MaxMinRelativeMarginFiller(Unit unit, double pstPenaltyCost, double negativeMarginObjectiveCoefficient, double ptdfSumLowerBound) {
+        super(unit, pstPenaltyCost);
         this.negativeMarginObjectiveCoefficient = negativeMarginObjectiveCoefficient;
         this.ptdfSumLowerBound = ptdfSumLowerBound;
     }
@@ -103,26 +102,6 @@ public class MaxMinRelativeMarginFiller extends MaxMinMarginFiller {
                 MPConstraint minimumMarginPositive = linearProblem.addMinimumRelativeMarginConstraint(-linearProblem.infinity(), maxFlow.get(), cnec, LinearProblem.MarginExtension.ABOVE_THRESHOLD);
                 minimumMarginPositive.setCoefficient(minRelMarginVariable, unitConversionCoefficient * relMarginCoef);
                 minimumMarginPositive.setCoefficient(flowVariable, 1);
-            }
-        });
-
-        // For CNECs of operators not sharing RAs, deactivate their participation in the definition of the minimum margin
-        // TODO : a large part of the code is common with MaxMinMarginFiller -> maybe merge the two fillers
-        double bigM = 2 * getLargestCnecThreshold(raoData);
-        getCnecsForOperatorsNotToOptimize(raoData).forEach(cnec -> {
-            MPVariable marginDecreaseBinaryVariable = linearProblem.getMarginDecreaseBinaryVariable(cnec);
-            if (marginDecreaseBinaryVariable == null) {
-                throw new FaraoException(String.format("Margin decrease binary variable has not yet been created for Cnec %s", cnec.getId()));
-            }
-            MPConstraint minimumMarginNegative = linearProblem.getMinimumRelativeMarginConstraint(cnec, LinearProblem.MarginExtension.BELOW_THRESHOLD);
-            if (minimumMarginNegative != null) {
-                minimumMarginNegative.setCoefficient(marginDecreaseBinaryVariable, bigM);
-                minimumMarginNegative.setUb(minimumMarginNegative.ub() + bigM);
-            }
-            MPConstraint minimumMarginPositive = linearProblem.getMinimumRelativeMarginConstraint(cnec, LinearProblem.MarginExtension.ABOVE_THRESHOLD);
-            if (minimumMarginPositive != null) {
-                minimumMarginPositive.setCoefficient(marginDecreaseBinaryVariable, bigM);
-                minimumMarginPositive.setUb(minimumMarginPositive.ub() + bigM);
             }
         });
     }
