+++ ra-optimisation/rao-commons/src/test/java/com/farao_community/farao/rao_commons/linear_optimisation/fillers/MaxMinRelativeMarginFillerTest.java
@@ -6,9 +6,7 @@
  */
 package com.farao_community.farao.rao_commons.linear_optimisation.fillers;
 
-import com.farao_community.farao.commons.Unit;
 import com.farao_community.farao.data.crac_api.cnec.BranchCnec;
-import com.farao_community.farao.data.crac_api.threshold.BranchThresholdRule;
 import com.farao_community.farao.data.crac_result_extensions.CnecResultExtension;
 import com.farao_community.farao.rao_commons.linear_optimisation.LinearProblem;
 import com.google.ortools.linearsolver.MPConstraint;
@@ -18,13 +16,12 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import java.util.Collections;
-
 import static com.farao_community.farao.commons.Unit.AMPERE;
 import static com.farao_community.farao.commons.Unit.MEGAWATT;
 import static com.farao_community.farao.rao_api.RaoParameters.DEFAULT_PST_PENALTY_COST;
 import static junit.framework.TestCase.assertTrue;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 /**
  * @author Peter Mitri {@literal <peter.mitri at rte-france.com>}
@@ -41,7 +38,7 @@ public class MaxMinRelativeMarginFillerTest extends AbstractFillerTest {
     public void setUp() {
         init();
         coreProblemFiller = new CoreProblemFiller();
-        maxMinRelativeMarginFiller = new MaxMinRelativeMarginFiller(MEGAWATT, DEFAULT_PST_PENALTY_COST, null, 1000, 0.01);
+        maxMinRelativeMarginFiller = new MaxMinRelativeMarginFiller(MEGAWATT, DEFAULT_PST_PENALTY_COST, 1000, 0.01);
     }
 
     private void fillProblemWithCoreFiller() {
@@ -153,86 +150,5 @@ public class MaxMinRelativeMarginFillerTest extends AbstractFillerTest {
         // check the number of variables and constraints
         assertEquals(5, linearProblem.getSolver().numVariables());
         assertEquals(7, linearProblem.getSolver().numConstraints());
-    }
-
-    private void setupOperatorsNotToOptimize() {
-        // Add a cnec
-        crac.newBranchCnec().setId("Line NL - N - preventive")
-                .newNetworkElement().setId("NNL1AA1  NNL2AA1  1").add()
-                .newThreshold().setRule(BranchThresholdRule.ON_LEFT_SIDE).setMax(800.0).setMin(-1000.).setUnit(Unit.MEGAWATT).add()
-                .optimized()
-                .setInstant(crac.getInstant("N"))
-                .setOperator("NL")
-                .add();
-        // Set initial margins on both preventive CNECs
-        cnecNl = crac.getBranchCnec("Line NL - N - preventive");
-        cnecFr = crac.getBranchCnec("Tieline BE FR - N - preventive");
-
-        // Create filler with new operatorsNotToOptimize and fill
-        maxMinRelativeMarginFiller = new MaxMinRelativeMarginFiller(MEGAWATT, DEFAULT_PST_PENALTY_COST, Collections.singleton("NL"), 1000, 0.01);
-    }
-
-    @Test
-    public void testCnecsNotToOptimizeBinaryVar() {
-        setupOperatorsNotToOptimize();
-        initRaoData(crac.getPreventiveState());
-        cnecFr.getExtension(CnecResultExtension.class).getVariant(raoData.getPreOptimVariantId()).setFlowInMW(600); // with a threshold of +750/-750
-        cnecNl.getExtension(CnecResultExtension.class).getVariant(raoData.getPreOptimVariantId()).setFlowInMW(400); // with a threshold of +800/-1000
-        double worstMarginDecreaseOnCnecNl = (800 - 400) - (750 - 600);
-
-        fillProblemWithCoreFiller();
-        maxMinRelativeMarginFiller.fill(raoData, linearProblem);
-
-        // Verify existence of margin_decrease binary variable
-        assertNull(linearProblem.getMarginDecreaseBinaryVariable(cnecFr));
-        MPVariable binaryVar = linearProblem.getMarginDecreaseBinaryVariable(cnecNl);
-        assertNotNull(binaryVar);
-
-        // Get flow variable
-        MPVariable flowVar = linearProblem.getFlowVariable(cnecNl);
-
-        // Verify existence of margin_decrease definition constraints
-        assertNull(linearProblem.getMarginDecreaseConstraint(cnecFr, LinearProblem.MarginExtension.BELOW_THRESHOLD));
-        assertNull(linearProblem.getMarginDecreaseConstraint(cnecFr, LinearProblem.MarginExtension.ABOVE_THRESHOLD));
-
-        MPConstraint marginDecreaseConstraintMin = linearProblem.getMarginDecreaseConstraint(cnecNl, LinearProblem.MarginExtension.BELOW_THRESHOLD);
-        assertNotNull(marginDecreaseConstraintMin);
-        assertEquals(linearProblem.infinity(), marginDecreaseConstraintMin.ub(), DOUBLE_TOLERANCE);
-        assertEquals(-1000 + (800 - 400), marginDecreaseConstraintMin.lb(), DOUBLE_TOLERANCE);
-        assertEquals(1.0, marginDecreaseConstraintMin.getCoefficient(flowVar), DOUBLE_TOLERANCE);
-        assertEquals(worstMarginDecreaseOnCnecNl, marginDecreaseConstraintMin.getCoefficient(binaryVar), DOUBLE_TOLERANCE);
-
-        MPConstraint marginDecreaseConstraintMax = linearProblem.getMarginDecreaseConstraint(cnecNl, LinearProblem.MarginExtension.ABOVE_THRESHOLD);
-        assertNotNull(marginDecreaseConstraintMax);
-        assertEquals(linearProblem.infinity(), marginDecreaseConstraintMax.ub(), DOUBLE_TOLERANCE);
-        assertEquals(-800 + (800 - 400), marginDecreaseConstraintMax.lb(), DOUBLE_TOLERANCE);
-        assertEquals(-1.0, marginDecreaseConstraintMax.getCoefficient(flowVar), DOUBLE_TOLERANCE);
-        assertEquals(worstMarginDecreaseOnCnecNl, marginDecreaseConstraintMax.getCoefficient(binaryVar), DOUBLE_TOLERANCE);
-    }
-
-    @Test
-    public void testExcludeCnecsNotToOptimizeInMinMargin() {
-        setupOperatorsNotToOptimize();
-        initRaoData(crac.getPreventiveState());
-        cnecFr.getExtension(CnecResultExtension.class).getVariant(raoData.getPreOptimVariantId()).setFlowInMW(600); // with a threshold of +750/-750
-        cnecNl.getExtension(CnecResultExtension.class).getVariant(raoData.getPreOptimVariantId()).setFlowInMW(400); // with a threshold of +800/-1000
-        double maxAbsThreshold = 1000;
-        fillProblemWithCoreFiller();
-        maxMinRelativeMarginFiller.fill(raoData, linearProblem);
-
-        // Test that cnecFr's constraint does not have a bigM
-        assertEquals(750.0, linearProblem.getMinimumMarginConstraint(cnecFr, LinearProblem.MarginExtension.BELOW_THRESHOLD).ub(), DOUBLE_TOLERANCE);
-        assertEquals(750.0, linearProblem.getMinimumMarginConstraint(cnecFr, LinearProblem.MarginExtension.ABOVE_THRESHOLD).ub(), DOUBLE_TOLERANCE);
-        assertEquals(750.0, linearProblem.getMinimumRelativeMarginConstraint(cnecFr, LinearProblem.MarginExtension.BELOW_THRESHOLD).ub(), DOUBLE_TOLERANCE);
-        assertEquals(750.0, linearProblem.getMinimumRelativeMarginConstraint(cnecFr, LinearProblem.MarginExtension.ABOVE_THRESHOLD).ub(), DOUBLE_TOLERANCE);
-
-        // Test that cnecNl's constraint does have a bigM
-        MPVariable marginDecreaseVariable = linearProblem.getMarginDecreaseBinaryVariable(cnecNl);
-        MPConstraint minMarginDefMin = linearProblem.getMinimumMarginConstraint(cnecNl, LinearProblem.MarginExtension.BELOW_THRESHOLD);
-        assertEquals(1000 + 2 * maxAbsThreshold, minMarginDefMin.ub(), DOUBLE_TOLERANCE);
-        assertEquals(2 * maxAbsThreshold, minMarginDefMin.getCoefficient(marginDecreaseVariable), DOUBLE_TOLERANCE);
-        MPConstraint minMarginDefMax = linearProblem.getMinimumMarginConstraint(cnecNl, LinearProblem.MarginExtension.ABOVE_THRESHOLD);
-        assertEquals(800 + 2 * maxAbsThreshold, minMarginDefMax.ub(), DOUBLE_TOLERANCE);
-        assertEquals(2 * maxAbsThreshold, minMarginDefMax.getCoefficient(marginDecreaseVariable), DOUBLE_TOLERANCE);
     }
 }
